requirements.txt
openai>=1.14.0
python-dotenv>=1.0.1
tenacity>=8.2.3
markdown>=3.5
